KSA-5  Hyun Simple api support IOS KSA-46  Hyun support Deeplink

+++ sample/src/mobileMain/kotlin/kim/jeonghyeon/sample/repository/WordRepository.kt
@@ -1,29 +1,34 @@
 package kim.jeonghyeon.sample.repository
 
-import kim.jeonghyeon.coroutine.repositoryFlow
 import kim.jeonghyeon.pergist.asListFlow
 import kim.jeonghyeon.sample.Word
 import kim.jeonghyeon.sample.WordQueries
 import kim.jeonghyeon.sample.api.SimpleApi
-import kim.jeonghyeon.type.ResourceFlow
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.flow
 
 interface WordRepository {
-    fun getWord(): ResourceFlow<List<Word>>
+    fun getWord(): Flow<List<Word>>
     suspend fun insertWord(word: String)
 }
 
 var fetchedWordApi = false
 
 class WordRepositoryImpl(val api: SimpleApi, val query: WordQueries) : WordRepository {
-    override fun getWord(): ResourceFlow<List<Word>> = repositoryFlow(
-        loadFromDb = { query.selectAll().asListFlow() },
-        shouldFetch = { !fetchedWordApi },//call first time after app started
-        callApi = { api.getWords().also { fetchedWordApi = true } },
-        saveResponse = {
+
+    override fun getWord(): Flow<List<Word>> = flow {
+        if (!fetchedWordApi) {
+            val response = api.getWords()
             query.deleteAll()
-            it.forEach { query.insert(it) }
+            response.forEach { query.insert(it) }
+            fetchedWordApi = true
         }
-    )
+
+        query.selectAll().asListFlow().collect {
+            emit(it)
+        }
+    }
 
     override suspend fun insertWord(word: String) {
         api.addWord(word)
